[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-	int A = 0, B = 0;
-	scanf("%d %d", &A, &B);
-	printf("%d", (A + B) * (A - B));
+	long A = 0, B = 0;
+	scanf("%ld %ld", &A, &B);
+	printf("%ld", (A + B) * (A - B));
 
 	return 0;
 }
